+++ common/src/main/java/net/satisfy/vinery/core/block/StemBlock.java
@@ -7,7 +7,6 @@ import net.minecraft.sounds.SoundEvents;
 import net.minecraft.sounds.SoundSource;
 import net.minecraft.util.Mth;
 import net.minecraft.util.RandomSource;
-import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResult;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.Item;
@@ -57,7 +56,6 @@ public abstract class StemBlock extends Block implements BonemealableBlock {
     }
 
     @Override
-    @SuppressWarnings("deprecation")
     public @NotNull InteractionResult useWithoutItem(BlockState state, Level world, BlockPos pos, Player player, BlockHitResult hit) {
         final int age = state.getValue(AGE);
         if (age > 3) {