+++ fabric/src/main/java/net/satisfy/vinery/platform/fabric/PlatformHelperImpl.java
@@ -18,8 +18,8 @@ import net.satisfy.vinery.platform.PlatformHelper;
 import java.util.List;
 import java.util.function.Supplier;
 
-public class PlatformHelperImpl extends PlatformHelper
-{
+public class PlatformHelperImpl extends PlatformHelper {
+
     public static int getTotalFermentationTime() {
         VineryFabricConfig config = AutoConfig.getConfigHolder(VineryFabricConfig.class).getConfig();
         return config.blocks.totalFermentationTime;
@@ -60,13 +60,17 @@ public class PlatformHelperImpl extends PlatformHelper
         return config.blocks.grapeGrowthChance;
     }
 
+    public static double getGrapeGrowthMultiplier() {
+        VineryFabricConfig config = AutoConfig.getConfigHolder(VineryFabricConfig.class).getConfig();
+        return config.blocks.grapeGrowthMultiplier;
+    }
+
     public static int getWineMaxLevel() {
         VineryFabricConfig config = AutoConfig.getConfigHolder(VineryFabricConfig.class).getConfig();
         return config.items.wine.maxLevel;
     }
 
     public static int getWineStartDuration() {
-
         VineryFabricConfig config = AutoConfig.getConfigHolder(VineryFabricConfig.class).getConfig();
         return config.items.wine.startDuration;
     }
@@ -101,11 +105,6 @@ public class PlatformHelperImpl extends PlatformHelper
         return config.items.banner.isShowTooltipEnabled();
     }
 
-    public static List<String> getBasketBlacklist() {
-        VineryFabricConfig config = AutoConfig.getConfigHolder(VineryFabricConfig.class).getConfig();
-        return config.items.basket.blacklist.basketBlacklist;
-    }
-
     public static double getTraderSpawnChance() {
         VineryFabricConfig config = AutoConfig.getConfigHolder(VineryFabricConfig.class).getConfig();
         return config.trader.spawnChance;
@@ -122,7 +121,11 @@ public class PlatformHelperImpl extends PlatformHelper
     }
 
     public static <T extends Entity> Supplier<EntityType<T>> registerBoatType(String name, EntityType.EntityFactory<T> factory, MobCategory category, float width, float height, int clientTrackingRange) {
-        EntityType<T> registry = Registry.register(BuiltInRegistries.ENTITY_TYPE, ResourceLocation.fromNamespaceAndPath(Vinery.MOD_ID, name), FabricEntityTypeBuilder.create(category, factory).dimensions(EntityDimensions.scalable(width, height)).trackRangeChunks(clientTrackingRange).build());
+        EntityType<T> registry = Registry.register(
+                BuiltInRegistries.ENTITY_TYPE,
+                ResourceLocation.fromNamespaceAndPath(Vinery.MOD_ID, name),
+                FabricEntityTypeBuilder.create(category, factory).dimensions(EntityDimensions.scalable(width, height)).trackRangeChunks(clientTrackingRange).build()
+        );
         return () -> registry;
     }
 }